# openApi.yaml
openapi: 3.0.3
info:
  title: Commulab System API
  description: >
    API untuk autentikasi, manajemen pasien, sesi konseling,
    chat AI, transkrip, evaluasi, dan laporan statistik.
  version: 1.0.0
servers:
  - url: http://localhost:3000/api
    description: Local development server
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: "Error message"
    AuthSuccessResponse:
      type: object
      properties:
        status:
          type: string
          example: ok
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
          description: JWT token berlaku 3 jam
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
    User:
      type: object
      properties:
        user_id:
          type: integer
          example: 1
        username:
          type: string
          example: "johndoe"
        email:
          type: string
          format: email
          example: "john@example.com"
        role:
          type: string
          enum: [user, admin]
          example: "user"
        is_admin:
          type: boolean
          example: false
        created_at:
          type: string
          format: date-time
          example: "2024-12-27T10:00:00.000Z"
...
security:
  - BearerAuth: []
paths:
  /auth/register:
    post:
      tags: [Auth]
      security: []
      summary: Registrasi user baru
      description: >
        Membuat akun user baru dengan username, email, dan password.
        Password harus minimal 8 karakter dan mengandung huruf besar, 
        huruf kecil, dan angka. Setelah registrasi berhasil, user 
        otomatis mendapat JWT token yang berlaku 3 jam.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, email, password]
              properties:
                username:
                  type: string
                  example: johndoe
                  description: Username unik untuk user
                email:
                  type: string
                  format: email
                  example: john@example.com
                  description: Email unik untuk user
                password:
                  type: string
                  minLength: 8
                  example: Password123
                  description: >
                    Minimal 8 karakter, harus mengandung:
                    - Minimal 1 huruf besar (A-Z)
                    - Minimal 1 huruf kecil (a-z)
                    - Minimal 1 angka (0-9)
            examples:
              valid:
                summary: Request valid
                value:
                  username: johndoe
                  email: john@example.com
                  password: Password123
      responses:
        "200":
          description: Registrasi berhasil
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthSuccessResponse"
              examples:
                success:
                  summary: Registrasi berhasil
                  value:
                    status: ok
                    user:
                      user_id: 1
                      username: johndoe
                      email: john@example.com
                      role: user
                      is_admin: false
                      created_at: "2024-12-27T10:00:00.000Z"
                    token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        "400":
          description: Registrasi gagal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                missing_fields:
                  summary: Field tidak lengkap
                  value:
                    status: error
                    message: Username, email, dan password harus diisi
                weak_password:
                  summary: Password tidak memenuhi syarat
                  value:
                    status: error
                    message: Password minimal 8 karakter, mengandung huruf besar, huruf kecil, dan angka
                email_exists:
                  summary: Email sudah terdaftar
                  value:
                    status: error
                    message: Email sudah terdaftar. Silakan login.
  /auth/login:
    post:
      tags: [Auth]
      security: []
      summary: Login user
      description: >
        Login menggunakan email dan password. Jika berhasil, 
        akan mendapat JWT token yang berlaku 3 jam. Token ini 
        digunakan untuk autentikasi endpoint lain dengan format 
        header: Authorization: Bearer <token>
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                  example: john@example.com
                password:
                  type: string
                  example: Password123
            examples:
              valid:
                summary: Login valid
                value:
                  email: john@example.com
                  password: Password123
      responses:
        "200":
          description: Login berhasil
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthSuccessResponse"
              examples:
                success_user:
                  summary: Login sebagai user biasa
                  value:
                    status: ok
                    user:
                      user_id: 1
                      username: johndoe
                      email: john@example.com
                      role: user
                      is_admin: false
                      created_at: "2024-12-27T10:00:00.000Z"
                    token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
                success_admin:
                  summary: Login sebagai admin
                  value:
                    status: ok
                    user:
                      user_id: 2
                      username: admin
                      email: admin@example.com
                      role: admin
                      is_admin: true
                      created_at: "2024-12-27T09:00:00.000Z"
                    token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        "400":
          description: Login gagal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                missing_fields:
                  summary: Field tidak lengkap
                  value:
                    status: error
                    message: Email dan password harus diisi
                invalid_credentials:
                  summary: Email atau password salah
                  value:
                    status: error
                    message: Email atau password salah.
  /chat:
    post:
      tags: [Chat]
      summary: Kirim pesan chat ke AI
      description: >
        Mengirim pesan user ke AI dan mendapat response dalam bentuk 
        array messages yang berisi text, facial expression, animation, 
        dan audio. Jika message kosong, AI akan memberikan greeting default.
        Prosody data opsional untuk analisis intonasi suara user.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                message:
                  type: string
                  description: Pesan dari user
                  example: "Saya merasa sedih hari ini"
                session_id:
                  type: integer
                  description: ID session konseling yang aktif
                  example: 1
                prosody_data:
                  type: object
                  description: Data prosody/intonasi suara (opsional)
                  properties:
                    pitch:
                      type: number
                      example: 220.5
                    energy:
                      type: number
                      example: 0.75
                    tempo:
                      type: number
                      example: 1.2
            examples:
              with_message:
                summary: Chat dengan pesan
                value:
                  message: "Saya merasa cemas belakangan ini"
                  session_id: 1
                  prosody_data:
                    pitch: 250.3
                    energy: 0.65
                    tempo: 1.1
              empty_message:
                summary: Greeting (message kosong)
                value:
                  message: ""
                  session_id: 1
      responses:
        "200":
          description: Respon chat berhasil
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  messages:
                    type: array
                    items:
                      type: object
                      properties:
                        text:
                          type: string
                          description: Text response dari AI
                          example: "Saya mengerti perasaanmu. Ceritakan lebih lanjut?"
                        facialExpression:
                          type: string
                          description: Ekspresi wajah avatar
                          enum: [smile, sad, angry, surprised, default]
                          example: "smile"
                        animation:
                          type: string
                          description: Nama animasi avatar
                          example: "Talking_1"
                        audio:
                          type: string
                          nullable: true
                          description: Base64 audio atau URL audio dari ElevenLabs
                          example: "data:audio/mpeg;base64,..."
              examples:
                ai_response:
                  summary: Response normal dari AI
                  value:
                    success: true
                    messages:
                      - text: "Saya mendengarkan. Ceritakan lebih lanjut tentang perasaanmu."
                        facialExpression: "smile"
                        animation: "Talking_1"
                        audio: "data:audio/mpeg;base64,..."
                default_greeting:
                  summary: Greeting default (message kosong)
                  value:
                    messages:
                      - text: "Hai... bagaimana kabarmu hari ini?"
                        facialExpression: "smile"
                        animation: "Talking_1"
                        audio: null
        "500":
          description: Error server atau API key belum diset
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: "API key belum diset."
              examples:
                api_key_error:
                  summary: API key tidak tersedia
                  value:
                    message: "API key belum diset."
                processing_error:
                  summary: Error saat processing
                  value:
                    success: false
                    message: "Error processing chat message"
  /chat/persona/{session_id}:
    get:
      tags: [Chat]
      summary: Ambil persona aktif berdasarkan session
      description: >
        Mengambil data persona AI yang sedang aktif untuk session tertentu.
        Persona berisi karakteristik AI seperti personality, background, 
        dan traits yang digunakan dalam percakapan.
      parameters:
        - name: session_id
          in: path
          required: true
          description: ID session konseling
          schema:
            type: integer
          example: 1
      responses:
        "200":
          description: Persona berhasil diambil
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  persona:
                    type: object
                    description: Data persona AI aktif
                    properties:
                      persona_id:
                        type: integer
                        example: 1
                      patient_id:
                        type: integer
                        example: 5
                      name:
                        type: string
                        example: "Budi Santoso"
                      background_story:
                        type: string
                        example: "Seorang mahasiswa yang mengalami kecemasan..."
                      personality_type:
                        type: string
                        example: "introvert"
                      symptom_intensity:
                        type: integer
                        example: 7
                      personality_traits:
                        type: object
                        example: { "anxiety_level": "high", "openness": "medium" }
              examples:
                success:
                  summary: Persona ditemukan
                  value:
                    success: true
                    persona:
                      persona_id: 1
                      patient_id: 5
                      name: "Budi Santoso"
                      background_story: "Mahasiswa semester 6 yang mengalami anxiety"
                      personality_type: "introvert"
                      symptom_intensity: 7
                      personality_traits:
                        anxiety_level: "high"
                        social_comfort: "low"
        "500":
          description: Error saat mengambil persona
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: "Persona tidak ditemukan untuk session ini"
  /chat/set-persona-from-patient:
    post:
      tags: [Chat]
      summary: Set persona AI dari data patient
      description: >
        Mengatur persona AI berdasarkan data pasien yang ada. 
        Persona akan digunakan AI untuk berperilaku sesuai 
        karakteristik pasien dalam simulasi konseling.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [patient_id]
              properties:
                patient_id:
                  type: integer
                  description: ID pasien yang akan dijadikan persona
                  example: 5
            examples:
              valid:
                summary: Set persona dengan patient_id
                value:
                  patient_id: 5
      responses:
        "200":
          description: Persona berhasil diset
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Persona berhasil diset dari data patient"
                  activePersona:
                    type: object
                    description: Data persona yang baru diset
                    properties:
                      persona_id:
                        type: integer
                      patient_id:
                        type: integer
                      name:
                        type: string
                      background_story:
                        type: string
                  patient_id:
                    type: integer
                    description: ID pasien yang digunakan
                    example: 5
              examples:
                success:
                  summary: Persona berhasil diset
                  value:
                    success: true
                    message: "Persona berhasil diset dari data patient"
                    activePersona:
                      persona_id: 1
                      patient_id: 5
                      name: "Budi Santoso"
                      background_story: "Mahasiswa yang mengalami anxiety"
                      personality_type: "introvert"
                      symptom_intensity: 7
                    patient_id: 5
        "400":
          description: patient_id tidak diberikan
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: "patient_id is required"
        "500":
          description: Error saat set persona
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: "Patient tidak ditemukan"
...